Moved HessianBuffer from subproblem solvers to ExactHessian (the only concerned)
