+++ uno/ingredients/subproblem_solvers/BQPD/BQPDQuadraticProgram.hpp
@@ -9,7 +9,6 @@
 #include "ingredients/subproblem_solvers/QuadraticProgram.hpp"
 #include "ingredients/subproblem_solvers/SolverWorkspace.hpp"
 #include "linear_algebra/Vector.hpp"
-#include "optimization/HessianBuffer.hpp"
 
 namespace uno {
    // forward declarations
@@ -76,7 +75,6 @@ namespace uno {
       Vector<uno_int> hessian_row_indices{};
       Vector<uno_int> hessian_column_indices{};
       Vector<double> hessian_values{}; // workspace
-      HessianBuffer hessian_buffer{}; // buffer used to evaluate the Hessian
       bool hessian_evaluation_required{false};
       mutable Vector<double> hessian_vector_product{};
