+++ uno/ingredients/subproblem_solvers/HiGHS/HiGHSQuadraticProgram.hpp
@@ -8,7 +8,6 @@
 #include "ingredients/subproblem_solvers/QuadraticProgram.hpp"
 #include "ingredients/subproblem_solvers/SolverWorkspace.hpp"
 #include "linear_algebra/Vector.hpp"
-#include "optimization/HessianBuffer.hpp"
 
 namespace uno {
    // forward declarations
@@ -51,7 +50,6 @@ namespace uno {
       Vector<uno_int> hessian_row_indices{};
       Vector<uno_int> hessian_column_indices{};
       Vector<double> hessian_values{}; // workspace
-      HessianBuffer hessian_buffer{}; // buffer used to evaluate the Hessian
       Vector<size_t> hessian_permutation_vector{};
 
    protected:
