+++ uno/ingredients/subproblem_solvers/WoodburyEQPSolver.hpp
@@ -8,7 +8,6 @@
 #include "DirectSymmetricIndefiniteLinearSolver.hpp"
 #include "SubproblemSolver.hpp"
 #include "optimization/Direction.hpp"
-#include "optimization/HessianBuffer.hpp"
 
 namespace uno {
    // forward declarations
@@ -56,7 +55,6 @@ namespace uno {
       const DirectQuasiNewtonHessian& hessian_model;
       std::unique_ptr<DirectSymmetricIndefiniteLinearSolver<double>> linear_solver;
       bool analysis_performed{false};
-      HessianBuffer hessian_buffer; // buffer used to evaluate the Hessian
 
       void compute_low_rank_correction(const Subproblem& subproblem, Vector<double>& b) const;
       [[nodiscard]] static bool solve_dense_indefinite_system(DenseMatrix<double>& T, const Vector<double>& c, Vector<double>& d);
